Fix(wasm): Resolve f64 to u32 migration errors
- Update HP and AC to u32 in Creature struct in model.rs.
- Fix all float literals and comparisons (e.g., > 0.0) to use integer equivalents for HP and AC.
- Cast u32 HP/AC values to f64 for scoring and averaging where floats are required.
- Cast f64 damage values to u32 when applying to HP.
- Use abs_diff for u32 absolute differences.
- Ensure all related logic in actions, targeting, aggregation, resolution, and validation files is consistent with integer HP/AC.

+++ simulation-wasm/src/combat_stats.rs
@@ -91,7 +91,7 @@ impl CombatantStats {
         
         Self {
             creature_id: creature.id.clone(),
-            ac: creature.ac,
+            ac: creature.ac as f64,
             dpr: total_dpr,
             hit_probability: if attack_count > 0 {
                 // Average hit probability across all attacks
